ai-fix: 구조화된 출력시 글자수 제한 로직 강화

+++ ProjecTree-Inference/app/api/schemas/nodes.py
@@ -16,6 +16,6 @@ class NodeCreateResponse(BaseModel):
     node_id: Optional[int] = Field(None, description="생성된 노드 ID")
     parent_id: Optional[int] = Field(None, description="부모 노드 ID")
     candidate_id: Optional[int] = Field(None, description="후보 노드 ID")
-    x_pos: Optional[int] = Field(None, description="노드 X 좌표")
-    y_pos: Optional[int] = Field(None, description="노드 Y 좌표")
+    x_pos: Optional[float] = Field(None, description="노드 X 좌표")
+    y_pos: Optional[float] = Field(None, description="노드 Y 좌표")
